add gen_index_page to gen_pages.py
